GCC on 64-bit Linux has 32-bit int and 64-bit size_t; not taking this into account led to a stack overwrite that broke most of the new decompression code on this platform.
SVN-Revision: 539

+++ libarchive/archive_read_support_compression_gzip.c
@@ -112,7 +112,7 @@ gzip_bidder_bid(struct archive_read_filter_bidder *self,
     struct archive_read_filter *filter)
 {
 	const unsigned char *buffer;
-	size_t avail;
+	ssize_t avail;
 	int bits_checked;
 
 	(void)self; /* UNUSED */
@@ -349,7 +349,7 @@ gzip_filter_read(struct archive_read_filter *self, const void **p)
 	struct private_data *state;
 	size_t read_avail, decompressed;
 	const void *read_buf;
-	int ret;
+	ssize_t ret;
 
 	state = (struct private_data *)self->data;
 	read_avail = 0;
