+++ libarchive/archive_read_support_compression_lzma.c
@@ -117,7 +117,7 @@ lzma_bidder_bid(struct archive_read_filter_bidder *self,
     struct archive_read_filter *filter)
 {
 	const unsigned char *buffer;
-	size_t avail;
+	ssize_t avail;
 	int bits_checked;
 
 	(void)self; /* UNUSED */
@@ -186,7 +186,7 @@ lzma_bidder_init(struct archive_read_filter *self)
 	void *out_block;
 	const char *buff;
 	struct private_data *state;
-	int ret;
+	ssize_t ret;
 
 	self->code = ARCHIVE_COMPRESSION_LZMA;
 	self->name = "lzma";
@@ -267,7 +267,7 @@ lzma_filter_read(struct archive_read_filter *self, const void **p)
 	struct private_data *state;
 	size_t read_avail, decompressed;
 	const void *read_buf;
-	int ret;
+	ssize_t ret;
 
 	state = (struct private_data *)self->data;
 	read_avail = 0;
